Updated: Azure PredictSoilHumidity function
Now works, but is missing proper commenting/documentation.

It takes a https request with a body like below, where target is the soilhumidity threshold we wish to know how long it takes to reach. The dummy model just returns the same number as an integer (i.e. 45 hours remaining for the below).

{
  "inputs": {
    "target": [45.7]
  }
}